pyproject.toml
[tool.poetry]
name = "stl2png"
version = "0.1.0"
description = ""
authors = ["Oren Mazor <oren.mazor@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
vtkplotlib = "^1.4.1"
numpy-stl = "^2.16.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
